backend/api/requirements.txt
python-dotenv
shortuuid
Flask
requests 
Flask-Cors==5.0.0
mysql-connector-python
PyJWT